+++ test/defaults/main.tf
@@ -31,5 +31,6 @@ module "gke" {
 }
 
 output "gke" {
-  value = module.gke
+  value     = module.gke
+  sensitive = true
 }