+++ test/overrides/main.tf
@@ -4,6 +4,8 @@ locals {
   environment = var.environment
 }
 
+data "google_project" "project" {}
+
 module "gke" {
   source = "../../"
 
@@ -21,7 +23,22 @@ module "gke" {
     "10.10.0.0/18" = "local-network"
   }
 
+  workload_identity_config = [{
+    identity_namespace = format("%s.svc.id.goog", data.google_project.project.project_id)
+  }]
+
   database_encryption_kms_key = "" # data.google_kms_crypto_key.shared.self_link
+
+  # Provide generic defaults for our nodepools so we don't have to set them for each of them seperately.
+  # Merged with the module's node_pool_defaults output to fill in missing variables
+  node_pool_defaults = merge(
+    module.gke.node_pool_defaults,
+    {
+      workload_metadata_config = [{
+        node_metadata = "GKE_METADATA_SERVER" # Enables workload identity on the node.
+      }]
+    }
+  )
 
   node_pools = {
     "primary" = {
@@ -43,5 +60,6 @@ module "gke" {
 }
 
 output "gke" {
-  value = module.gke
+  value     = module.gke
+  sensitive = true
 }