+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## V1.1.0 (2021-02-08)
+
+This release requires Terraform 0.14 or higher
+
+### Features
+* Added support for workload identity configuration
+* Added support for management configuration (auto_repair and auto_upgrade settings)
+
 ## v1.0.0 (2020-11-23)
 
 * Initial release